refactor(scanner): extract strategy/direction.py — resolve_direction_params + metrics + ATR aliases (#225 PR2)
